feat: params_manager (WIP)
@lorenzo-bianchi
I'll explain in a voice message, hopefully before leaving but I'm in a run.
I'll also explain the (minor) modifications required for your Python script to be integrated in this.
I decided to do this right now because as soon as I start working on updating the Flight Stack, I'll either use this or waste an astonishing amount of time.