feat: dua_qos
@alessandrotenaglia @lorenzo-bianchi @giorgiomanca97
This is the common source of all our QoS profiles.
It's a shared library so as long as one process loads it, all running nodes will be able to access it.
As with everything in dua_utils, you can contribute if you feel like this needs new features or some other kind of updates.
See the README and header for more info.