feat: dua_qos: get_action_server/client_options
@lorenzo-bianchi @alessandrotenaglia @giorgiomanca97
So, when you create an action server/client, you usually want to specify a callback group for it.
Because of the APIs' signature, if you want to specify a callback group, you also have to specify server/client options.
Server/client options are largely made of QoS profiles which we should never modify, which means... boilerplate code!
This eliminates the need to copy and paste stuff everywhere. Just pass the result of a get_*() call to the create_* function.